requirements.txt
setuptools
python-dotenv